requirements.txt
fastapi
uvicorn
litellm
python-dotenv
langfuse==2.60.8
httpx
rich
pandas
numpy
scikit-learn
judgy
rank-bm25
tqdm
pydantic
typing-extensions
matplotlib
seaborn
plotly
